pixtral attention to temporary patches and remove compiler replacements
