+++ unsloth_zoo/temporary_patches/__init__.py
@@ -20,3 +20,4 @@ from .gemma import *
 from .misc import *
 from .gemma3n import *
 from .gpt_oss import *
+from .pixtral import *
